profilePublications. Improve mistakes
